Version 1.2 Update Colum in Output.csv
PowerShell Script: File Deletion
Overview
This PowerShell script reads a list of file paths from an input text file, deletes the files and folders specified in the paths, and exports a CSV file containing information about the deleted files.

Prerequisites
Create an input text file named input.txt and place it in C:\Users\Public\.
In the input.txt file, enter the file paths you want to delete, with one path per line.
Usage
To run the script, open PowerShell and execute the script file. After the script has finished running, it will generate an Outputs.csv file in the C:\Users\Public\ folder.

Detailed Description
The script performs the following actions:

Reads the paths from the input.txt file.
Loops through each path in the input file.
Checks if the path exists.
If the path exists, the script calculates the file size and gets the list of files to be deleted.
Deletes the files and folders specified in the path.
Appends the deletion details to an array.
If the path doesn't exist, the script adds an entry with "File doesn't exist" for that path.
Calculates the script execution time.
Exports the file deletion details to a CSV file named Outputs.csv.
Output
The output CSV file (Outputs.csv) contains the following columns:

Files Deleted: Shows "Yes" if files have been deleted, and "File doesn't exist" if the path is invalid.
File Path: The path of the file or folder specified in the input file.
File Size: The total file size (in MB) of the deleted files.
Deleted Files: A list of deleted files, separated by new lines.
Additionally, the last row of the output CSV file contains the script execution time.

Note
When taking screenshots for the documentation, you can capture images of:

The input text file (input.txt) with the list of file paths.
The PowerShell console where you execute the script.
The output CSV file (Outputs.csv) with the deletion details.
Remember to add these images to your documentation to provide visual context for the steps and output.